# Make rb_gc_impl_writebarrier_remember Ractor-safe