rb_gc_impl_writebarrier_remember is not Ractor safe because it writes to bitmaps and also pushes onto the mark stack during incremental marking. We should acquire the VM lock to prevent race conditions.

In the case that the object is not old, there is no performance impact. However, we can see a performance impact in this microbenchmark where the object is old:

```ruby
4.times.map do
  Ractor.new do
    ary = []

    3.times { GC.start }

    10_000_000.times do |i|
      ary.push(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17)
      ary.clear
    end
  end
end.map(&:value)
```

Before:

    Time (mean ± σ):     682.4 ms ±   5.1 ms    [User: 2564.8 ms, System: 16.0 ms]

After:

    Time (mean ± σ):      5.522 s ±  0.096 s    [User: 8.237 s, System: 7.931 s]